minilide/requirements.txt
pandas
fpdf
python-dotenv
matplotlib
nicegui
requests
beautifulsoup4
pushbullet.py